// src/interfaces/meetings.interface.ts
export interface Meetings {
    meeting_id: string;
    date_scheduled: string;
    meeting_name: string;
    staff_id: string;
    room_id: string;
    hub_id: string;
    check_in_count: number;
    check_out_count: number;
    training_material: string;
    meeting_start_time: string;
    meeting_end_time: string;    
    meeting_type: string;
    imei: string;
    app_version: string;
  }
  